+++ doc/examples/briefcase__main__.py
@@ -1,4 +1,0 @@
-from appname.app import main
-
-if __name__ == '__main__':
-    main()